+++ components/DrawerItem.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { StyleSheet } from "react-native";
+import { StyleSheet, TouchableOpacity } from "react-native";
 import { Block, Text, theme } from "galio-framework";
 
 import Icon from "./Icon";
@@ -15,7 +15,7 @@ class DrawerItem extends React.Component {
           <Icon
             name="shop"
             family="ArgonExtra"
-            size={10}
+            size={14}
             color={focused ? "white" : argonTheme.COLORS.PRIMARY}
           />
         );
@@ -24,7 +24,7 @@ class DrawerItem extends React.Component {
           <Icon
             name="map-big"
             family="ArgonExtra"
-            size={12}
+            size={14}
             color={focused ? "white" : argonTheme.COLORS.ERROR}
           />
         );
@@ -33,7 +33,7 @@ class DrawerItem extends React.Component {
           <Icon
             name="spaceship"
             family="ArgonExtra"
-            size={12}
+            size={14}
             color={focused ? "white" : argonTheme.COLORS.PRIMARY}
           />
         );
@@ -42,7 +42,7 @@ class DrawerItem extends React.Component {
           <Icon
             name="chart-pie-35"
             family="ArgonExtra"
-            size={12}
+            size={14}
             color={focused ? "white" : argonTheme.COLORS.WARNING}
           />
         );
@@ -51,12 +51,17 @@ class DrawerItem extends React.Component {
           <Icon
             name="calendar-date"
             family="ArgonExtra"
-            size={12}
+            size={14}
             color={focused ? "white" : argonTheme.COLORS.INFO}
           />
         );
       case "Getting Started":
-        return <Icon />;
+        return (<Icon
+          name="spaceship"
+          family="ArgonExtra"
+          size={14}
+          color={focused ? "white" : "rgba(0,0,0,0.5)"}
+        />);
       case "Log out":
         return <Icon />;
       default:
@@ -65,7 +70,7 @@ class DrawerItem extends React.Component {
   };
 
   render() {
-    const { focused, title } = this.props;
+    const { focused, title, navigation } = this.props;
 
     const containerStyles = [
       styles.defaultStyle,
@@ -73,28 +78,39 @@ class DrawerItem extends React.Component {
     ];
 
     return (
-      <Block flex row style={containerStyles}>
-        <Block middle flex={0.1} style={{ marginRight: 5 }}>
-          {this.renderIcon()}
+      <TouchableOpacity
+        style={{ height: 60 }}
+        onPress={() =>
+          title == "Getting Started"
+            ? Linking.openURL(
+                "https://demos.creative-tim.com/argon-pro-react-native/docs/"
+              ).catch(err => console.error("An error occurred", err))
+            : navigation.navigate(title)
+        }
+      >
+        <Block flex row style={containerStyles}>
+          <Block middle flex={0.1} style={{ marginRight: 5 }}>
+            {this.renderIcon()}
+          </Block>
+          <Block row center flex={0.9}>
+            <Text
+              size={15}
+              bold={focused ? true : false}
+              color={focused ? "white" : "rgba(0,0,0,0.5)"}
+            >
+              {title}
+            </Text>
+          </Block>
         </Block>
-        <Block row center flex={0.9}>
-          <Text
-            size={15}
-            bold={focused ? true : false}
-            color={focused ? "white" : "rgba(0,0,0,0.5)"}
-          >
-            {title}
-          </Text>
-        </Block>
-      </Block>
+      </TouchableOpacity>
     );
   }
 }
 
 const styles = StyleSheet.create({
   defaultStyle: {
-    paddingVertical: 15,
-    paddingHorizontal: 14
+    paddingVertical: 16,
+    paddingHorizontal: 16
   },
   activeStyle: {
     backgroundColor: argonTheme.COLORS.ACTIVE,